fix: Correctly initialize context in inspector client init()
(Taken from the Android PR. Original commit message:)
I'm not sure about this, but I think the code was initializing context_
with its own initial value (i.e. nullptr), rather than the isolate's
context. This seemed to trigger an error when sending a message to the
frontend, which this change fixed.