Track "super" references in ObjCDataWrappers using traceable structs
Previously extended class constructors stored their CacheItem in a single
InternalField, and the CacheItem was retained by the Caches collection for the
Isolate.

Consider adding a separate wrapper type, or a flag to the ObjCDataWrapper struct,
to indicate the disposal semantics --- However so far I don't think this is an issue.