refactor Struct types
This is one of the larger changes to the codebase --- As usual, there is a refactoring
related to things such as the use of v8::Externals, which are now replaced by JSObjects
with 2 embedder slots suitable for holding CPPGC values and tracing them effectively.

The old approach to retaining the "parent" struct of a StructWrapper is simplified to
rely on a traced value, avoiding the need for shared pointer tracking.